refactor(auth): extract privileged login
